+++ engine/ref/soft/r_main.c
@@ -563,7 +563,7 @@ static void R_DrawEntitiesOnList( void )
 
 	if( !FBitSet( RI.rvp.flags, RF_ONLY_CLIENTDRAW ))
 	{
-		gEngfuncs.CL_DrawEFX( tr.frametime, false );
+		gEngfuncs.CL_DrawEFX( tr.frametime, false, false );
 	}
 
 	if( FBitSet( RI.rvp.flags, RF_DRAW_WORLD ))
@@ -618,7 +618,7 @@ static void R_DrawEntitiesOnList( void )
 	if( !FBitSet( RI.rvp.flags, RF_ONLY_CLIENTDRAW ))
 	{
 		R_AllowFog( false );
-		gEngfuncs.CL_DrawEFX( tr.frametime, true );
+		gEngfuncs.CL_DrawEFX( tr.frametime, true, false );
 		R_AllowFog( true );
 	}