+++ engine/ref/soft/r_part.c
@@ -45,7 +45,7 @@ CL_DrawParticles
 update particle color, position, free expired and draw it
 ================
 */
-void GAME_EXPORT CL_DrawParticles( double frametime, particle_t *cl_active_particles, float partsize )
+void GAME_EXPORT CL_DrawParticles( double frametime, particle_t *cl_active_particles, float partsize, qboolean draw_only )
 {
 	vec3_t right, up;
 
@@ -81,8 +81,10 @@ void GAME_EXPORT CL_DrawParticles( double frametime, particle_t *cl_active_parti
 			VectorScale( RI.cull_vright, size, right );
 			VectorScale( RI.cull_vup, size, up );
 
-			p->color = bound( 0, p->color, 255 );
-			color24 color = tr.palette[p->color];
+			short color_index = (short)bound( 0, p->color, 255 );
+			if( !draw_only )
+				p->color = color_index;
+			color24 color = tr.palette[color_index];
 
 			int alpha = 255 * ( p->die - gp_cl->time ) * 16.0f;
 			if( alpha > 255 || p->type == pt_static )
@@ -107,7 +109,8 @@ void GAME_EXPORT CL_DrawParticles( double frametime, particle_t *cl_active_parti
 			r_stats.c_particle_count++;
 		}
 
-		gEngfuncs.CL_ThinkParticle( frametime, p );
+		if( !draw_only )
+			gEngfuncs.CL_ThinkParticle( frametime, p );
 	}
 
 	TriEnd();
@@ -134,13 +137,13 @@ CL_DrawTracers
 update tracer color, position, free expired and draw it
 ================
 */
-void GAME_EXPORT CL_DrawTracers( double frametime, particle_t *cl_active_tracers )
+void GAME_EXPORT CL_DrawTracers( double frametime, particle_t *cl_active_tracers, qboolean draw_only )
 {
 	vec3_t screenLast, screen;
 	vec3_t start, end, delta;
 
 	// update tracer color if this is changed
-	if( FBitSet( tracerred->flags | tracergreen->flags | tracerblue->flags | traceralpha->flags, FCVAR_CHANGED ))
+	if( !draw_only && FBitSet( tracerred->flags | tracergreen->flags | tracerblue->flags | traceralpha->flags, FCVAR_CHANGED ))
 	{
 		color24 *customColors = &gTracerColors[4];
 		customColors->r = (byte)( tracerred->value * traceralpha->value * 255 );
@@ -230,22 +233,25 @@ void GAME_EXPORT CL_DrawTracers( double frametime, particle_t *cl_active_tracers
 			TriEnd();
 		}
 
-		// evaluate position
-		VectorMA( p->org, frametime, p->vel, p->org );
-
-		if( p->type == pt_grav )
+		if( !draw_only )
 		{
-			p->vel[0] *= scale;
-			p->vel[1] *= scale;
-			p->vel[2] -= gravity;
+			// evaluate position
+			VectorMA( p->org, frametime, p->vel, p->org );
 
-			p->unused = 255 * ( p->die - gp_cl->time ) * 2;
-			if( p->unused > 255 )
-				p->unused = 255;
-		}
-		else if( p->type == pt_slowgrav )
-		{
-			p->vel[2] = gravity * 0.05;
+			if( p->type == pt_grav )
+			{
+				p->vel[0] *= scale;
+				p->vel[1] *= scale;
+				p->vel[2] -= gravity;
+
+				p->unused = 255 * ( p->die - gp_cl->time ) * 2;
+				if( p->unused > 255 )
+					p->unused = 255;
+			}
+			else if( p->type == pt_slowgrav )
+			{
+				p->vel[2] = gravity * 0.05;
+			}
 		}
 	}
 
@@ -270,8 +276,22 @@ void GAME_EXPORT CL_DrawParticlesExternal( const ref_viewpass_t *rvp, qboolean t
 	memcpy( RI.visbytes, tr.visbytes, gpGlobals->visbytes );
 	tr.frametime = frametime;
 
-	gEngfuncs.CL_DrawEFX( frametime, trans_pass );
+	gEngfuncs.CL_DrawEFX( frametime, trans_pass, false );
 
 	// restore internal state
+	RI = oldRI;
+}
+
+void GAME_EXPORT CL_DrawEFXView( const ref_viewpass_t *rvp, qboolean trans_pass, qboolean draw_only )
+{
+	ref_instance_t oldRI = RI;
+
+	if( !rvp )
+		return;
+
+	R_SetupRefParams( rvp );
+	R_SetupFrustum();
+	memcpy( RI.visbytes, tr.visbytes, gpGlobals->visbytes );
+	gEngfuncs.CL_DrawEFX( tr.frametime, trans_pass, draw_only );
 	RI = oldRI;
 }